fix: current agent name exist

+++ apps/mobile/src/hooks/perps/usePerpsState.ts
@@ -343,7 +343,10 @@ export const usePerpsState = () => {
         isSameAddress(agent.address, agentAddress),
       );
       if (!item) {
-        if (extraAgents.length >= 3) {
+        const existAgentName = extraAgents.find(
+          agent => agent.name === PERPS_AGENT_NAME,
+        );
+        if (!existAgentName && extraAgents.length >= 3) {
           // 超过3个，需要删除一个
           deleteAgentCbRef.current = async () => {
             const deleteItem = minBy(extraAgents, agent => agent.validUntil);
